TravisCI & Karma runner (issue #4)
Unfortunately, the webpack-dev-server isn't flexible enough for launching tests.
Especially, it is tricky to use this runner in a CI service. Therefore, the Karma runner
(https://karma-runner.github.io/1.0/index.html) replaced the webpack-dev-server
as more flexible and configurable test runner. Currently, it is configured to launch locally tests
in the Chrome.

Besides introducing the Karma runner, the TravisCI got set up to launch tests in Firefox (it
is shiped by the TravisCI, therefore, there won't be extra time spent on intalling a browser).