don't bother to the engine in case of refreshing a page (#19)
When a user refreshes the page, the client mustn't make a request to the engine,
because it leads to increasing transition counts between the referrer and
the current page. Also, it is wrong from the performance perspective, we can reuse the already made prediction.

The session storage is a good solution for caching data per tab. Data doesn't leak to other tabs.