A new public api to listen to making predictions (#12)
The end users may like to use different tracking systems in order to track accuracy of predictions made by the sirko.
This change introduces a new api to the end user. The end user can provide a callback which gets called once the prediction is made.
The callback receives the current prediction and an argument which means either the previous prediction was correct or not.

Internally, there are a few changes related to internal callbacks. As a chain of callbacks is painful, promises got
used to make code straightforward.