a fallback for browsers which don't support the prerender hint (#22)
The fallback is based on a Service Worker which fetches and caches the predicted page.
Once the user visits that page, the response gets served from the cache. It doesn't provide 100%
functionality of the prerender hint, for example, it doesn't execute JS and doesn't load
assets etc. But, the page gets served faster anyway.