prefetch assets for the predicted page (#24)
Now, before requesting a prediction from the engine, the client
gathers information about assets (CSS, JavaScript) on the current
page. The gathered list gets sent along with the current path.
So, the engine is able to keep paths of pages and assets to them.
When the prediction is received from the engine, it contains
the path and assets of a next page. Thus, the client instructs
browsers to prefech the predicted page and assets to it.