migrate to Rollup (https://rollupjs.org/)
Rollup creates smaller bundles than webpack.

To use the Node.js's assert lib, more dependencies must've
been added. The Node.js's assert lib is designed to work
on the server side, it is a bit tricky to use it in the browser.
So, to avoid expanding a list of dependencies,
Chai (http://chaijs.com/) replaced the Node.js's assert lib.