Pipeline instead of processors (#25)
Every step which does something for the prediction model
is a pipe. Every pipe can modify given data, interrupt execution,
or return a promise object so a next pipe has to wait for it.

This refactoring made code more straitforward.
Even the code requesting the prediction from a server
got represented as pipe.