don't track transitions which modify data (#1)
Also fixes: sirko-io/engine#10, sirko-io/engine#15

When a user navigates between pages, there might be POST/PUT/DELETE
requests. Those requests modify state, so a prefetched page might be
stale.

Example

  GET /signin
  POST /signin
  GET /index

If the transition between the signin and index pages got tracked,
the index page would be prefetched when the user opens the signin page.
This behavior's been leaded to a number of issues.

To avoid the described issue, the service worker
stores all made requests on a page then the client part
asks the service worker to check whether there was a POST/PUT/DELETE
request between the referrer and the current page. If so,
the transition isn't tracked (the referrer parameter gets removed, thus,
the engine doesn't track the transition), but the engine makes a prediction
for the current page anyway. The transition is about the previous
and the current page, the prediction is about the next page.