offline support (#29)
The engine got changed to return a list of pages which can
be prefetched. Since they are already in the cache, now
instead of removing them they get shifted to another
cache (`sirko-offline`) and served when a user is offline.

Currently, the logic of serving pages looks like this.

As a first step the lib checks whether a requested resource
(a page or an asset) is in the `sirko-prefetch` cache,
if so, it gets served, otherwise it gets fetched and served.
If the fetch fails, we assume that the user is offline, thus,
the lib looks for the resource in the `sirko-offline` cache
and serves it from there.

The `sirko-prefetch` cache is only used to serve resources for
a next transition as it's been worked before. It gets cleared
when the users moves to another page.

The `sirko-offline` cache accumulates all prefetched resources
which might be served in the offline mode. This cache gets updated
with fresh resources as the user navigates. So, currently,
the lib doesn't clear them, it only overrides them over time.