don't cut paths of URLs which get sent to the engine (#33)
A described proposal for refactoring in #32 was taken
as a solution to fix the bug. It means the client only
considers pages covered by the secure connection (https).
It is fine because service workers only work for
the secure connection.

Also, this change fixed an issue described in #30. Thus, URLs
of assets leading to internal scripts of browsers aren't sent
to the engine, so, the client doesn't try to precache them.